[ACCEPTED] linear scan
(MY NOTES)
- First of all, you need to obtain the prices directly from the formula. You have no other choice.
- But then, how can we find maximum difference between a_i, a_j (i < j)? First solution is to check every pair's difference. But that leads to O(n^2) and exceeds to time limit. What is the better way?

- In this example: {133, 81, 38, 42, 141, 54, 91, 149, 39, 37}
we first take 133 as max_element and calculate the max_diff with every next element. But when we reach to an element bigger than 133 (141), we update the max_element (= 141) and continue calculating the max_diff with 141. Why we don't continue with that 133? Because after, if we would reach to an element a_k which 133 - a_k > max_diff, then certainly 141 - a_k > 133 - a_k > max_diff. So, why continue with 133 when we can have 141?!